\name{pystr-package}
\alias{pystr}
\title{Python String Methods in R}
\description{
  String operations the Python way - a package for those of us who miss Python's
  string methods while we're working in R.
}
\details{
  \tabular{ll}{
    Package: \tab pystr\cr
    Version: \tab 1.0.0\cr
    Date: \tab 2015-07-05\cr
  }
}
\author{
  Nicole White

  For support, please create a new GitHub issue:
  \url{https://github.com/nicolewhite/pystr/issues}
}
