% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pystr.R
\name{pystr}
\alias{pystr}
\title{Python String Methods in R}
\description{
String operations the Python way - a package for those of us who miss Python's
string methods while we're working in R. For support, please
\href{https://github.com/nicolewhite/pystr/issues}{create a new GitHub issue}.
}
\author{
Nicole White
}

