% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pystr_capitalize.R
\name{pystr_capitalize}
\alias{pystr_capitalize}
\title{Capitalize a string.}
\usage{
pystr_capitalize(str)
}
\arguments{
\item{str}{A string.}
}
\value{
A string.
}
\description{
Return a copy of the string with its first character capitalized and the rest lowercased.
}
\examples{
pystr_capitalize("ONCE UPON A TIME, ")
}
\references{
\url{https://docs.python.org/3/library/stdtypes.html#str.capitalize}
}
\seealso{
\code{\link{pystr_title}}
}

