% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pystr_index.R
\name{pystr_index}
\alias{pystr_index}
\title{Find the lowest index of a substring.}
\usage{
pystr_index(str, sub)
}
\arguments{
\item{str}{A string.}

\item{sub}{A string.}
}
\value{
An integer.
}
\description{
Like \code{\link{pystr_find}} but raises an error if \code{sub} is not found.
}
\examples{
pystr_index("abcxyzabc", "abc")
\dontrun{
pystr_index("abcxyzabc", "123")
}
}
\references{
\url{https://docs.python.org/3/library/stdtypes.html#str.index}
}
\seealso{
\code{\link{pystr_rindex}}
}

