% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pystr_isalpha.R
\name{pystr_isalpha}
\alias{pystr_isalpha}
\title{Check if a string is alphabetic.}
\usage{
pystr_isalpha(str)
}
\arguments{
\item{str}{A string.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Return \code{TRUE} if all characters in the string are alphabetic and there is at least one character,
\code{FALSE} otherwise.
}
\examples{
pystr_isalpha("abc")
pystr_isalpha("abc123")
pystr_isalpha("abc!")
}
\references{
\url{https://docs.python.org/3/library/stdtypes.html#str.isalpha}
}
\seealso{
\code{\link{pystr_isalnum}}, \code{\link{pystr_isnumeric}}
}

