% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pystr_isnumeric.R
\name{pystr_isnumeric}
\alias{pystr_isnumeric}
\title{Check if a string is numeric.}
\usage{
pystr_isnumeric(str)
}
\arguments{
\item{str}{A string.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Return \code{TRUE} if all characters in the string are numeric characters, and there is at least one
character, \code{FALSE} otherwise.
}
\examples{
pystr_isnumeric("123")
pystr_isnumeric("123a")
pystr_isnumeric("123!")
}
\references{
\url{https://docs.python.org/3/library/stdtypes.html#str.isnumeric}
}
\seealso{
\code{\link{pystr_isalpha}}, \code{\link{pystr_isalnum}}
}

