% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pystr_isspace.R
\name{pystr_isspace}
\alias{pystr_isspace}
\title{Check if a string is whitespace.}
\usage{
pystr_isspace(str)
}
\arguments{
\item{str}{A string.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Return \code{TRUE} if there are only whitespace characters in the string and there is at least
one character, \code{FALSE} otherwise.
}
\examples{
pystr_isspace("    ")
pystr_isspace("  a ")
}
\references{
\url{https://docs.python.org/3/library/stdtypes.html#str.isspace}
}

