% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pystr_istitle.R
\name{pystr_istitle}
\alias{pystr_istitle}
\title{Check if a string is titlecase.}
\usage{
pystr_istitle(str)
}
\arguments{
\item{str}{A character vector.}
}
\value{
A logical vector.
}
\description{
Return \code{TRUE} if the string is a titlecased string and there is at least one
character, for example uppercase characters may only follow uncased characters and lowercase
characters only cased ones. Return \code{FALSE} otherwise.
}
\examples{
pystr_istitle("I Am A Title")
pystr_istitle("I Am not A Title")
}
\references{
\url{https://docs.python.org/3/library/stdtypes.html#str.istitle}
}

