% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pystr_isupper.R
\name{pystr_isupper}
\alias{pystr_isupper}
\title{Check if a string is uppercase.}
\usage{
pystr_isupper(str)
}
\arguments{
\item{str}{A string.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Return \code{TRUE} if all cased characters in the string are uppercase and there is at
least one cased character, \code{FALSE} otherwise.
}
\examples{
pystr_islower("ALL UPPERCASE!")
pystr_islower("All Uppercase?")
pystr_islower("ABC123")
}
\references{
\url{https://docs.python.org/3/library/stdtypes.html#str.isupper}
}
\seealso{
\code{\link{pystr_islower}}
}

