% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pystr_join.R
\name{pystr_join}
\alias{pystr_join}
\title{Join the elements of a vector or list into a string.}
\usage{
pystr_join(iterable, sep = "")
}
\arguments{
\item{iterable}{A vector or list.}

\item{sep}{A string.}
}
\value{
A string.
}
\description{
Return a string which is the concatenation of the elements in the iterable \code{iterable},
where \code{sep} is the separator between elements.
}
\examples{
pystr_join(c("A", "B", "C"))
pystr_join(c(1, 2, 3), "+")
pystr_join(list("Apples", "Oranges"), ", ")
}
\references{
\url{https://docs.python.org/3/library/stdtypes.html#str.join}
}
\seealso{
\code{\link{pystr_split}}
}

