% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pystr_lower.R
\name{pystr_lower}
\alias{pystr_lower}
\title{Lowercase a string.}
\usage{
pystr_lower(str)
}
\arguments{
\item{str}{A character vector.}
}
\value{
A character vector.
}
\description{
Return a copy of the string with all the cased characters converted to lowercase.
}
\examples{
pystr_lower("LOWERCASE ME!")
}
\references{
\url{https://docs.python.org/3/library/stdtypes.html#str.lower}
}
\seealso{
\code{\link{pystr_upper}}
}

