% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pystr_makestrans.R
\name{pystr_maketrans}
\alias{pystr_maketrans}
\title{Create a character map.}
\usage{
pystr_maketrans(x, y)
}
\arguments{
\item{x}{A string.}

\item{y}{A string.}
}
\value{
A list of character mappings for use in \code{\link{pystr_translate}}.
}
\description{
Create a list of character mappings usable for \code{\link{pystr_translate}}.
}
\details{
The two arguments must be strings of equal length, and in the resulting
dictionary, each character in \code{x} will be mapped to the character at the
same position in \code{y}.
}
\examples{
map = pystr_maketrans("abc", "123")
pystr_translate("a blue cat", map)
}
\references{
\url{https://docs.python.org/3/library/stdtypes.html#str.maketrans}
}
\seealso{
\code{\link{pystr_translate}}
}

