% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pystr_partition.R
\name{pystr_partition}
\alias{pystr_partition}
\title{Partition a string.}
\usage{
pystr_partition(str, sep)
}
\arguments{
\item{str}{A character vector.}

\item{sep}{A character string.}
}
\value{
A list of character vectors.
}
\description{
Split the string at the first occurrence of \code{sep}, and return a list of character vectors containing the part
before the separator, the separator itself, and the part after the separator.
}
\details{
If the separator is not found, return a character vector containing the string itself, followed by two empty strings.
}
\examples{
pystr_partition("onetwothreeonetwothree", "two")
}
\references{
\url{https://docs.python.org/3/library/stdtypes.html#str.partition}
}
\seealso{
\code{\link{pystr_rpartition}}
}

