% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pystr_rpartition.R
\name{pystr_rpartition}
\alias{pystr_rpartition}
\title{Partition a string from the right.}
\usage{
pystr_rpartition(str, sep)
}
\arguments{
\item{str}{A character vector.}

\item{sep}{A character string.}
}
\value{
A character vector.
}
\description{
Split the string at the last occurrence of \code{sep}, and return a list of character vectors containing the part before the separator,
the separator itself, and the part after the separator.
}
\details{
If the separator is not found, return a character vector containing two empty strings, followed by the string itself.
}
\examples{
pystr_rpartition("onetwothreeonetwothree", "two")
}
\references{
\url{https://docs.python.org/3/library/stdtypes.html#str.rpartition}
}
\seealso{
\code{\link{pystr_partition}}
}

