% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pystr_rstrip.R
\name{pystr_rstrip}
\alias{pystr_rstrip}
\title{Right strip a string.}
\usage{
pystr_rstrip(str, chars = " ")
}
\arguments{
\item{str}{A string.}

\item{chars}{A string.}
}
\value{
A string.
}
\description{
Return a copy of the string with trailing characters removed.
}
\details{
The \code{chars} argument is a string specifying the set of characters to be removed.
If omitted, the \code{chars} argument defaults to removing whitespace. The \code{chars} argument
is not a suffix; rather, all combinations of its values are stripped.
}
\examples{
pystr_rstrip("    spacious     ")
pystr_rstrip("www.example.com", ".omc")
}
\references{
\url{https://docs.python.org/3/library/stdtypes.html#str.rstrip}
}
\seealso{
\code{\link{pystr_lstrip}}
}

