% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pystr_swapcase.R
\name{pystr_swapcase}
\alias{pystr_swapcase}
\title{Swap the case of a string.}
\usage{
pystr_swapcase(str)
}
\arguments{
\item{str}{A string.}
}
\value{
A string.
}
\description{
Return a copy of the string with uppercase characters converted to lowercase and vice versa.
}
\examples{
pystr_swapcase("Swap Me!")
}
\references{
\url{https://docs.python.org/3/library/stdtypes.html#str.swapcase}
}

