% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pystr_title.R
\name{pystr_title}
\alias{pystr_title}
\title{Titlecase a string.}
\usage{
pystr_title(str)
}
\arguments{
\item{str}{A string.}
}
\value{
A string.
}
\description{
Return a titlecased version of the string where words start with an uppercase character
and the remaining characters are lowercase.
}
\examples{
pystr_title("make me pretty!")
}
\references{
\url{https://docs.python.org/3/library/stdtypes.html#str.title}
}

