% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pystr_upper.R
\name{pystr_upper}
\alias{pystr_upper}
\title{Uppercase a string.}
\usage{
pystr_upper(str)
}
\arguments{
\item{str}{A character vector.}
}
\value{
A character vector.
}
\description{
Return a copy of the string with all the cased characters converted to uppercase.
}
\examples{
pystr_upper("uppercase me!")
}
\references{
\url{https://docs.python.org/3/library/stdtypes.html#str.upper}
}
\seealso{
\code{\link{pystr_lower}}
}

