% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pystr_isalnum.R
\name{pystr_isalnum}
\alias{pystr_isalnum}
\title{Check if a string is alphanumeric.}
\usage{
pystr_isalnum(str)
}
\arguments{
\item{str}{A string.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Return \code{TRUE} if all characters in the string are alphanumeric and there is at
least one character, \code{FALSE} otherwise.
}
\examples{
pystr_isalnum("abc")
pystr_isalnum("abc123")
pystr_isalnum("abc123!")
}
\references{
\url{https://docs.python.org/3/library/stdtypes.html#str.isalnum}
}
\seealso{
\code{\link{pystr_isalpha}}, \code{\link{pystr_isnumeric}}
}

