% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pystr_islower.R
\name{pystr_islower}
\alias{pystr_islower}
\title{Check if a string is lowercase.}
\usage{
pystr_islower(str)
}
\arguments{
\item{str}{A character vector.}
}
\value{
A logical vector.
}
\description{
Return \code{TRUE} if all cased characters in the string are lowercase and there is at least one
cased character, \code{FALSE} otherwise.
}
\examples{
pystr_islower("all lowercase!")
pystr_islower("All Lowercase?")
pystr_islower("abc123")
}
\references{
\url{https://docs.python.org/3/library/stdtypes.html#str.islower}
}
\seealso{
\code{\link{pystr_isupper}}
}

