% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pystr_lstrip.R
\name{pystr_lstrip}
\alias{pystr_lstrip}
\title{Left strip a string.}
\usage{
pystr_lstrip(str, chars = " ")
}
\arguments{
\item{str}{A character vector.}

\item{chars}{A character string.}
}
\value{
A character vector.
}
\description{
Return a copy of the string with leading characters removed.
}
\details{
The \code{chars} argument is a string specifying the set of characters to be removed.
If omitted, the \code{chars} argument defaults to removing whitespace.
The \code{chars} argument is not a prefix; rather, all combinations of its values are stripped.
}
\examples{
pystr_lstrip("     spacious    ")
pystr_lstrip("www.example.com", "w.")
}
\references{
\url{https://docs.python.org/3/library/stdtypes.html#str.lstrip}
}
\seealso{
\code{\link{pystr_rstrip}}
}

